LayoutTests/proximity/create-event-expected.txt

Tests that document.createEvent() works with DeviceProximityEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof event == 'object' is true
PASS 'type' in event is true
PASS 'bubbles' in event is true
PASS 'cancelable' in event is true
PASS 'value' in event is true
PASS 'min' in event is true
PASS 'max' in event is true
PASS typeof event.type == 'string' is true
PASS typeof event.bubbles == 'boolean' is true
PASS typeof event.cancelable == 'boolean' is true
PASS typeof event.value == 'number' is true
PASS typeof event.min == 'number' is true
PASS typeof event.max == 'number' is true
PASS successfullyParsed is true

TEST COMPLETE
